rosotaLerobot/pyproject.toml
[project]
name = "rosota-lerobot"
version = "0.1.0"
description = "Minimal SO-100/SO-101 robot control system"
authors = [{ name = "Rosota", email = "rosota@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "pynput>=1.7.6",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "feetech-servo-sdk>=1.0.0",  # Required for FeetechMotorsBus
]

[project.scripts]
rosota-lerobot = "rosota_lerobot.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

